+++ src/main/java/shop/tryit/domain/question/QuestionRepository.java
@@ -3,6 +3,8 @@ package shop.tryit.domain.question;
 import java.util.Optional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import shop.tryit.domain.question.dto.QuestionSearchCondition;
+import shop.tryit.domain.question.dto.QuestionSearchDto;
 
 public interface QuestionRepository {
