+++ src/main/java/shop/tryit/domain/question/dto/QuestionSearchCondition.java
@@ -1,4 +1,4 @@
-package shop.tryit.domain.question;
+package shop.tryit.domain.question.dto;
 
 import lombok.Builder;
 import lombok.Getter;
