+++ src/main/java/shop/tryit/domain/question/dto/QuestionSearchDto.java
@@ -1,4 +1,4 @@
-package shop.tryit.domain.question;
+package shop.tryit.domain.question.dto;
 
 import com.querydsl.core.annotations.QueryProjection;
 import java.time.LocalDateTime;
