+++ src/main/java/shop/tryit/domain/question/service/QnAFacade.java
@@ -12,11 +12,11 @@ import shop.tryit.domain.answer.service.AnswerService;
 import shop.tryit.domain.member.Member;
 import shop.tryit.domain.member.MemberService;
 import shop.tryit.domain.question.Question;
-import shop.tryit.domain.question.QuestionSearchCondition;
-import shop.tryit.domain.question.QuestionSearchDto;
 import shop.tryit.domain.question.dto.QuestionCheckPasswordFormDto;
 import shop.tryit.domain.question.dto.QuestionFormDto;
 import shop.tryit.domain.question.dto.QuestionSaveFormDto;
+import shop.tryit.domain.question.dto.QuestionSearchCondition;
+import shop.tryit.domain.question.dto.QuestionSearchDto;
 
 @Component
 @RequiredArgsConstructor
