+++ src/main/java/shop/tryit/domain/question/service/QuestionService.java
@@ -9,8 +9,8 @@ import org.springframework.transaction.annotation.Transactional;
 import shop.tryit.domain.answer.AnswerRepository;
 import shop.tryit.domain.question.Question;
 import shop.tryit.domain.question.QuestionRepository;
-import shop.tryit.domain.question.QuestionSearchCondition;
-import shop.tryit.domain.question.QuestionSearchDto;
+import shop.tryit.domain.question.dto.QuestionSearchCondition;
+import shop.tryit.domain.question.dto.QuestionSearchDto;
 
 @Service
 @RequiredArgsConstructor
