+++ src/main/java/shop/tryit/repository/question/QuestionRepositoryImpl.java
@@ -7,8 +7,8 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Repository;
 import shop.tryit.domain.question.Question;
 import shop.tryit.domain.question.QuestionRepository;
-import shop.tryit.domain.question.QuestionSearchCondition;
-import shop.tryit.domain.question.QuestionSearchDto;
+import shop.tryit.domain.question.dto.QuestionSearchCondition;
+import shop.tryit.domain.question.dto.QuestionSearchDto;
 
 @Repository
 @RequiredArgsConstructor
