+++ src/test/java/shop/tryit/domain/question/QuestionServiceTests.java
@@ -17,6 +17,8 @@ import shop.tryit.domain.answer.Answer;
 import shop.tryit.domain.answer.AnswerRepository;
 import shop.tryit.domain.member.Member;
 import shop.tryit.domain.member.MemberRepository;
+import shop.tryit.domain.question.dto.QuestionSearchCondition;
+import shop.tryit.domain.question.dto.QuestionSearchDto;
 import shop.tryit.domain.question.service.QuestionService;
 import shop.tryit.repository.question.QuestionJpaRepository;
