+++ src/test/java/shop/tryit/repository/question/QuestionRepositoryImplTests.java
@@ -16,8 +16,8 @@ import shop.tryit.domain.member.Member;
 import shop.tryit.domain.member.MemberRepository;
 import shop.tryit.domain.question.Question;
 import shop.tryit.domain.question.QuestionRepository;
-import shop.tryit.domain.question.QuestionSearchCondition;
-import shop.tryit.domain.question.QuestionSearchDto;
+import shop.tryit.domain.question.dto.QuestionSearchCondition;
+import shop.tryit.domain.question.dto.QuestionSearchDto;
 
 @SpringBootTest
 @Transactional
